requirements.txt
absl-py==1.2.0
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
fastdist==1.1.3
Flask==2.2.2
fonttools==4.37.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0
gunicorn==20.1.0
idna==3.4
img2vec-pytorch==1.0.1
importlib-metadata==4.12.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
r2c==0.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-to-curl==1.0.7
rsa==4.9
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
